S2_NDRE_Monthly
The NDRE (Normalized Difference Red Edge) index is a vegetation index that is particularly useful for monitoring the health and vigor of plants. It is similar to the NDVI (Normalized Difference Vegetation Index), but instead of using the red band, it uses the red edge band, which is sensitive to changes in chlorophyll content.