S2_LAI
The LAI (Leaf Area Index) is a critical biophysical parameter used to quantify the amount of leaf material in a given area. It is defined as the one-sided green leaf area per unit ground area and is expressed in square meters of leaf area per square meter of ground area (m²/m²).