Landsat8_MSAVI
The MSAVI (Modified Soil Adjusted Vegetation Index) is a vegetation index used in remote sensing to minimize the effects of soil brightness when monitoring vegetation health and density. It is an enhancement of the SAVI (Soil Adjusted Vegetation Index) and is particularly useful in areas with sparse vegetation where soil influence is significant.