Create Gap_Fill_Landsat7_ETM+
The "striping error" in Landsat 7 ETM+ (Enhanced Thematic Mapper Plus) refers to a data artifact caused by the failure of the Scan Line Corrector (SLC) on the Landsat 7 satellite. This failure occurred on May 31, 2003, and has resulted in images with data gaps and stripes ever since.
Use this script to mitigate this error.